1st: fixed PR comments
